Created UAssetInsightWidget; Started binding widgets
Started creating UAssetInsightWidget which is a subclass of UUserWidget. This class is going to be the encapsulation of the asset insights.

I started messing with meta=(BindWidget) with test data. Was able to bind a function to a button inside cpp file, and trigger changes inside custom bound widgets using custom functions.

UI is a mess, I intend to tidy this up in a later commit.